streamlit-cloud-run/requirements.txt
streamlit==1.28.0
google-cloud-storage==2.13.0
six==1.16.0
